Add some helper macros for form fields
The goal is to make it easier to convert between django BoundFields and
NHS.UK design system component calls.

Normally in Django, the rendering to HTML of form fields is handled by a
combination of
- the field's Widget, and its associated template
- templates built into django itself, e.g. https://github.com/django/django/blob/main/django/forms/templates/django/forms/field.html

I explored the possibility of customising these various templates so
that we could call {{ form }} or {{ form.field }} in our templates, and
have it render the right components automatically. However, ultimately I
decided this was too complicated, because Django's widgets are quite
limited (they do not normally have access to the errors for example),
and the components contain extra functionality that is not
representable in Django forms, such as the conditional param in
checkboxes/radios.

That said, there is still a lot of boilerplate code in our templates, so
I think extracting some macros is still useful, even if we can't use
them for everything.