Expand participants app to include ScreeningEvent and Appointment
These models are used to capture information about a participant
and their screenings. I added them to the clinics app as part of a spike
and didn't really consider where they belong.

I think it makes more sense to include this in the participants app
and narrow the scope of the clinics app.

This inverts the dependency between clinics and participants.
participants will contain the Participant model and everything that
relates to a participant.

Migrating models between apps is fairly tricky due to each app having
its own set of migrations.