Rename "record_a_mammogram" to "mammograms"
This name is quite verbose, so we're shortening it to mammograms.

Colloquially this is refered to as a screening clinic, but our clinics
app contains the pages that a receptionist would use, whereas
mammograms contains the ones used by the mammographer in a first
appointment where the mammogram is done.

We are deliberately not naming this "appointments" because there could
be multiple appointments per screening episode, and we may add
additional journeys later on.

I've also moved the templates into the app.