Participant details page
Add email and phone number to the participant details card, and remove
risk level, to bring it in line with the prototype.

The new page also shows the NHS number and gender. I'm ignoring the SX
Number from the prototype as we are not going to need that unless we
deploy alongside NBSS.