Set the back link
This includes a skipped system test for navigating between appointments
and participant details. This is there because I considered
conditionally varying the back link in this case, but decided to keep it
simple for now.

However, we still want to implement this later, and even if we don't
implement it we will need a very similar test for navigating between
this and /participants (which doesn't yet exist) - so to save rewriting
all the steps I've kept the test.

We have a ticket to build a way of tracking users navigation through a
multi-step journey,

    origin -> page -> page -> result
           <-      <-      <-

When this is in place we will have a generic way of conditionally rendering back
links based on where the user has come from, and we can use that for
this back link as well.