Migrate to azure passwordless login
This approach is based on the one mentioned here:
https://github.com/Azure-Samples/msdocs-django-web-app-managed-identity/issues/19

However, I've added the extra step to avoiding refetching the token if
it hasn't already expired.
This step might be unnecessary, but I couldn't discern from the Azure
identity documentation whether the client already does something like
this internally.

This solution also assumes we're using the default "persistent
connections". Out of the box these refresh on every request, which we
might want to tweak when we are ready to do some load testing.
I noticed there is a new option since Django 5.1 to use the connection
pooling from pyscopg_pool[^1], but I don't see a way of making this
work with this passwordless login mechanism, because it passes
in the connection information once when the pool is initialised.

[^1]: https://www.psycopg.org/psycopg3/docs/api/pool.html