Try out multi-ecosystem-groups
This is a beta feature, so is subject to change, but we can try it out
with nhsuk-frontend / nhsuk-frontend-jinja, which need to be updated
together.

If this works well we could potentially further consolidate some of the
npm/pip groups, but it seems a bit tricky when you want some things
(like framework dependencies) to be a single PR and others (like linters
and other tools that have less effect on the app running in production)
to be grouped.