refactor: extract appointment status to an append only log
Introduce an AppointmentStatus table, which lets us give every status
change a timestamp.

To make it easy to access the current status, I'm using
@cached_property to extract it from the `statuses` related field.
This does not get reloaded automatically when `refresh_from_db` is
called, but it can be force reloaded with the fields argument[^1].

Filtering based on the latest status is more awkward with this schema,
as we need to use a subquery. However, this can be encapsulated in the
AppointmentQuerySet.

[^1]: https://docs.djangoproject.com/en/5.2/ref/models/instances/#django.db.models.Model.refresh_from_db