perf: don't fetch last known mammogram in AppointmentPresenter
This side effect causes an N+1 query on the clinics show page, because
we pass every appointment to the AppointmentPresenter.

Seeing as that page doesn't even show the last known mammogram, this
is entirely unnecessary.

In general, we should avoid making extra queries from within presenters,
as their job is to present data, not fetch it. Where that is the case,
we can reuse the presenters in different context without worrying
about extra overhead.