fix: eliminate n+1 problem with statuses
We've extracted statuses to a separate table, such that
an appointment can have many statuses. This is modelled in
the ORM by the `statuses` relation.

Previously we were calling `first()` to get the current one, however,
when you call first() on a queryset, django ORM always issues a new
query, even if there are objects in the prefetch cache[^1].  This
creates an N+1 query problem, because Appointment.current_status is
called in a loop when rendering the clinic show page.

As a workaround, we can prefetch all the statuses, and then select the
first one in python.

This is quite ugly but I couldn't find a neater
way to have 1-many relationship while just making use of the latest one.

It is possible to customise what gets fetched by passing a `Prefetch`
object when calling `prefetch_related`, but I felt this was overly
complicated for what we need now.