fix: ensure jinja templates are namespaced
We are using APP_DIRS=True so that templates are stored alongside the
apps, however, Django doesn't automatically namespace these templates,
so we need to explcitly include a subdirectory for the app in the directory
structure.

This is as per the Django tutorial:

> Now we might be able to get away with putting our templates directly in polls/templates (rather than creating another polls subdirectory), but it would actually be a bad idea. Django will choose the first template it finds whose name matches, and if you had a template with the same name in a different application, Django would be unable to distinguish between them. We need to be able to point Django at the right one, and the best way to ensure this is by namespacing them. That is, by putting those templates inside another directory named for the application itself.

https://docs.djangoproject.com/en/5.2/intro/tutorial03/#write-views-that-actually-do-something:~:text=Template%20namespacing,the%20application%20itself.

This fixes a conflict between mammograms/show.jinja and
participants/show.jinja which broke the participant record. This wasn't
caught by the system test because of another bug: it wasn't named
correctly so wasn't getting picked up by the suite.